Crytals
Pink Slimes get more powerful in the area of crystals.
