+++ Assets/Scripts/Crystal Action.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5597faf4aaf2dc049be33a6000a11c20
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
