+++ Assets/Scripts/Enemy Action/EnemyMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 79595aecb0df9834aa6e5e83f6040fc5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
