+++ Assets/Scripts/Enemy Action/EnemyShooting.cs
@@ -7,8 +7,14 @@ public class EnemyShooting : MonoBehaviour
     public GameObject bullet;
     public Transform bulletPos;
 
-    private float timer;
+    private float fireCount;
     public float fireTimer;
+    private float reloadCount;
+    private int ammoCount;
+    public int ammoMax;
+    public float reloadTimer;
+    private bool outAmmo = false;
+
     public bool shootOn = true;
 
     // Start is called before the first frame update
@@ -20,13 +26,22 @@ public class EnemyShooting : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        timer += Time.deltaTime;
+        fireCount += Time.deltaTime;
+        reloadCount += Time.deltaTime;
 
-        if(timer > fireTimer && GetComponent<EnemyLife>().ClosePlayer() && shootOn){
-            timer = 0;
+        if(fireCount > fireTimer && GetComponent<EnemyMovement>().ClosePlayer() && shootOn && !outAmmo){
+            fireCount = 0;
+            ammoCount++;
             shoot();
         }
-        
+        if(ammoCount > ammoMax){
+            outAmmo = true;
+            reloadCount = 0;
+        }
+        if(reloadCount > reloadTimer){
+            ammoCount = 0;
+            outAmmo = false;
+        }
         void shoot(){
             Instantiate(bullet, bulletPos.position, Quaternion.identity);
         }
