+++ Assets/Scripts/Player Action.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d65488bdbb58c6c47b1ab2cf5f51edff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
